Move Ascii helper to org.safere
