+++ safere/src/main/java/org/safere/Ascii.java
@@ -3,70 +3,70 @@
 // Modifications and Java port Copyright (c) 2026 Eddie Aftandilian.
 // Licensed under the BSD 3-Clause License (see LICENSE file).
 
-package org.safere.internal;
+package org.safere;
 
 /** ASCII character classification and case-folding utilities. */
-public final class Ascii {
+final class Ascii {
 
   private Ascii() {}
 
   /** Converts an ASCII uppercase character to lowercase; leaves other characters unchanged. */
-  public static char toLowerCase(char ch) {
+  static char toLowerCase(char ch) {
     return ch >= 'A' && ch <= 'Z' ? (char) (ch + 32) : ch;
   }
 
   /** Converts an ASCII lowercase character to uppercase; leaves other characters unchanged. */
-  public static char toUpperCase(char ch) {
+  static char toUpperCase(char ch) {
     return ch >= 'a' && ch <= 'z' ? (char) (ch - 32) : ch;
   }
 
   /** Converts an ASCII uppercase code point to lowercase; leaves other code points unchanged. */
-  public static int toLowerCase(int cp) {
+  static int toLowerCase(int cp) {
     return cp >= 'A' && cp <= 'Z' ? (cp + 32) : cp;
   }
 
   /** Converts an ASCII lowercase code point to uppercase; leaves other code points unchanged. */
-  public static int toUpperCase(int cp) {
+  static int toUpperCase(int cp) {
     return cp >= 'a' && cp <= 'z' ? (cp - 32) : cp;
   }
 
   /** Returns whether two characters are equal ignoring ASCII case. */
-  public static boolean equalsIgnoreCase(char a, char b) {
+  static boolean equalsIgnoreCase(char a, char b) {
     return a == b || toLowerCase(a) == toLowerCase(b);
   }
 
   /** Returns true if the code point is an ASCII uppercase letter. */
-  public static boolean isUpper(int r) {
+  static boolean isUpper(int r) {
     return r >= 'A' && r <= 'Z';
   }
 
   /** Returns true if the code point is an ASCII lowercase letter. */
-  public static boolean isLower(int r) {
+  static boolean isLower(int r) {
     return r >= 'a' && r <= 'z';
   }
 
   /** Returns true if the code point is an ASCII letter. */
-  public static boolean isAlpha(int r) {
+  static boolean isAlpha(int r) {
     return (r >= 'A' && r <= 'Z') || (r >= 'a' && r <= 'z');
   }
 
   /** Returns true if the code point is an ASCII digit. */
-  public static boolean isDigit(int r) {
+  static boolean isDigit(int r) {
     return r >= '0' && r <= '9';
   }
 
   /** Returns true if the code point is an ASCII letter or digit. */
-  public static boolean isAlnum(int r) {
+  static boolean isAlnum(int r) {
     return (r >= '0' && r <= '9') || (r >= 'A' && r <= 'Z') || (r >= 'a' && r <= 'z');
   }
 
   /** Returns true if the code point is an ASCII word character (letter, digit, or underscore). */
-  public static boolean isWordChar(int r) {
+  static boolean isWordChar(int r) {
     return isAlnum(r) || r == '_';
   }
 
   /** Returns true if the code point is an ASCII hex digit. */
-  public static boolean isHexDigit(int r) {
+  static boolean isHexDigit(int r) {
     return (r >= '0' && r <= '9') || (r >= 'A' && r <= 'F') || (r >= 'a' && r <= 'f');
   }
 
@@ -76,7 +76,7 @@ public final class Ascii {
    * @param r a code point
    * @return 0-15 for valid hex digits, -1 otherwise
    */
-  public static int unhex(int r) {
+  static int unhex(int r) {
     if (r >= '0' && r <= '9') {
       return r - '0';
     }
